Make some headway on gen_single_block() implementation

+++ yjit/src/yjit.rs
@@ -19,7 +19,7 @@ pub extern "C" fn rb_yjit_init() {
 /// Called from C code to begin compiling a function
 /// NOTE: this should be wrapped in RB_VM_LOCK_ENTER(), rb_vm_barrier() on the C side
 #[no_mangle]
-pub extern "C" fn rb_yjit_iseq_gen_entry_point(iseq: IseqPtr, insn_idx: usize, ec: EcPtr) -> *const u8 {
+pub extern "C" fn rb_yjit_iseq_gen_entry_point(iseq: IseqPtr, insn_idx: u32, ec: EcPtr) -> *const u8 {
 
     let maybe_code_ptr = crate::core::gen_entry_point(iseq, insn_idx, ec);
